refactor(SplashViewController): Правки кода без исправления ошибок или добавления новых функций
Проведен рефакторинг контроллера, логика работы вынесена в отдельную сущность (Presenter)

META: sprint_13,  ImageFeed
